Sticky Generator & Brainstorm Board
When you enter a # of lessons it populates the brainstorm board with
that # of divs (this still needs to be styled).

Sticky Generator clones sticky template when you click the cloning
icon. Stickies are draggable.

ISSUES - Something funky is happening when you clone the sticky more
than once. There is a double clone and the draggable function is not
working on the second sticky. Not sure what’s happening here.